move vote discussions to publikator
