+++ pages/vote/genossenschaft/diskutieren.js
@@ -1,4 +1,4 @@
-import ElectionDiscussionPage from '../../../components/Vote/ElectionDiscussionPage'
+import ElectionDiscussionPage from '../../../components/Vote/Legacy/ElectionDiscussionPage'
 import compose from 'lodash/flowRight'
 import { enforceMembership } from '../../../components/Auth/withMembership'
 import withMe from '../../../lib/apollo/withMe'
